- 1 screen has been added - A Drawer Widget has been added

+++ lib/src/models/item_model.dart
@@ -81,7 +81,7 @@ class _Result {
 
   String get title => _title;
 
-  double get vote_average => _vote_average;
+  double get vote_average => _vote_average.toDouble();
 
   bool get video => _video;
